requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
mpmath==1.1.0
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
powerlaw==1.4.6
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
snap-stanford==5.0.0
tweepy==3.8.0
urllib3==1.25.8
